+++ src/config/firebaseConfig.js
@@ -7,7 +7,6 @@ import 'firebase/functions';
 
 import { emulators } from '../../firebase.json';
 
-const functionsEmulator = `http://localhost:${emulators.functions.port}`;
 const firestoreEmulator = {
   host: `localhost:${emulators.firestore.port}`,
   ssl: false,
@@ -40,7 +39,7 @@ const { serverTimestamp, arrayRemove } = firebase.firestore.FieldValue;
 
 if (process.env.NODE_ENV === 'development' || window.Cypress) {
   db.settings(firestoreEmulator);
-  functions.useFunctionsEmulator(functionsEmulator);
+  functions.useEmulator('localhost', emulators.functions.port);
   console.log('Established connection to Firestore emulators');
 } else {
   console.log('Established connection to Firestore server');