+++ tests/e2e/.eslintrc.js
@@ -4,7 +4,4 @@ module.exports = {
     mocha: true,
     'cypress/globals': true,
   },
-  rules: {
-    strict: 'off',
-  },
 };