+++ tests/e2e/specs/01 Login/02 login_admin_user_spec.js
@@ -1,7 +1,6 @@
-/* eslint-disable */
 describe('Test Admin Login', () => {
   before(() => {
-    cy.visit('/').wait(3000);
+    cy.visit('/').wait(2000);
 
     cy.get('body').then($body => {
       if ($body.text().includes('Test Admin') || $body.text().includes(Cypress.env('VUE_APP_TESTUSER_USER'))) {